// src/app/interfaces/cliente.ts
export interface Cliente {
    nombre: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    rfc: string;
    mesComprobante: string;
    identificacionUrl: string;
    comprobanteDomicilioUrl: string;
    huellaIzquierdaUrl: string;
    huellaDerechaUrl: string;
    fotoUrl: string;


}